Fix sqlserver file loads
