+++ lib/composite_primary_keys.rb
@@ -99,6 +99,7 @@ require 'composite_primary_keys/nested_attributes'
 require 'composite_primary_keys/connection_adapters/abstract_adapter'
 require 'composite_primary_keys/connection_adapters/abstract_mysql_adapter'
 require 'composite_primary_keys/connection_adapters/postgresql/database_statements'
+require 'composite_primary_keys/connection_adapters/sqlserver/database_statements'
 
 require 'composite_primary_keys/relation/batches'
 require 'composite_primary_keys/relation/where_clause'
