+++ lib/composite_primary_keys/arel/sqlserver.rb
@@ -1,3 +1,5 @@
+require 'activerecord-sqlserver-adapter'
+
 module Arel
   module Visitors
     class SQLServer
